% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\docType{data}
\name{categories}
\alias{categories}
\title{Google Trends categories.}
\format{
A data frame with 1426 rows and 2 variables
}
\usage{
data("categories")
}
\description{
\itemize{
  \item name  Names of the categories
  \item id IDs of the categories
}
}
\keyword{datasets}
