\name{countries}
\alias{countries}
\docType{data}
\title{List of country codes supported by Google Trends}
\description{
A data set of countries and country codes.
}
\usage{data(countries)}
\format{
  A data frame with 245 observations on the following 2 variables.
  \describe{
    \item{\code{CODE}}{a character vector with values \code{AC}
      \code{AD} \code{AE} \code{AF} \code{AG} \code{AI} \code{AL}
      \code{AM} \code{AN} \code{AO} \code{AQ} \code{AR} \code{AS}
      \code{AT} \code{AU} \code{AW} \code{AZ} \code{BA} \code{BB}
      \code{BD} \code{BE} \code{BF} \code{BG} \code{BH} \code{BI}
      \code{BJ} \code{BM} \code{BN} \code{BO} \code{BR} \code{BS}
      \code{BT} \code{BV} \code{BW} \code{BY} \code{BZ} \code{CA}
      \code{CC} \code{CF} \code{CG} \code{CH} \code{CI} \code{CK}
      \code{CL} \code{CM} \code{CN} \code{CO} \code{CR} \code{CU}
      \code{CV} \code{CX} \code{CY} \code{CZ} \code{DE} \code{DJ}
      \code{DK} \code{DM} \code{DO} \code{DZ} \code{EC} \code{EE}
      \code{EG} \code{EH} \code{ER} \code{ES} \code{ET} \code{FI}
      \code{FJ} \code{FK} \code{FM} \code{FO} \code{FR} \code{FX}
      \code{GA} \code{GD} \code{GE} \code{GF} \code{GG} \code{GH}
      \code{GI} \code{GL} \code{GM} \code{GN} \code{GP} \code{GQ}
      \code{GR} \code{GS} \code{GT} \code{GU} \code{GW} \code{GY}
      \code{HK} \code{HM} \code{HN} \code{HR} \code{HT} \code{HU}
      \code{ID} \code{IE} \code{IL} \code{IM} \code{IN} \code{IO}
      \code{IQ} \code{IR} \code{IS} \code{IT} \code{JE} \code{JM}
      \code{JO} \code{JP} \code{KE} \code{KG} \code{KH} \code{KI}
      \code{KM} \code{KN} \code{KP} \code{KR} \code{KW} \code{KY}
      \code{KZ} \code{LA} \code{LB} \code{LC} \code{LI} \code{LK}
      \code{LR} \code{LS} \code{LT} \code{LU} \code{LV} \code{LY}
      \code{MA} \code{MC} \code{MD} \code{ME} \code{MG} \code{MH}
      \code{MK} \code{ML} \code{MM} \code{MN} \code{MO} \code{MP}
      \code{MQ} \code{MR} \code{MS} \code{MT} \code{MU} \code{MV}
      \code{MW} \code{MX} \code{MY} \code{MZ} \code{NC} \code{NE}
      \code{NF} \code{NG} \code{NI} \code{NL} \code{NO} \code{NP}
      \code{NR} \code{NU} \code{NZ} \code{OM} \code{PA} \code{PE}
      \code{PF} \code{PG} \code{PH} \code{PK} \code{PL} \code{PM}
      \code{PN} \code{PR} \code{PT} \code{PW} \code{PY} \code{QA}
      \code{RE} \code{RO} \code{RS} \code{RU} \code{RW} \code{SA}
      \code{SB} \code{SC} \code{SD} \code{SE} \code{SG} \code{SH}
      \code{SI} \code{SJ} \code{SK} \code{SL} \code{SM} \code{SN}
      \code{SO} \code{SR} \code{ST} \code{SV} \code{SY} \code{SZ}
      \code{TC} \code{TD} \code{TF} \code{TG} \code{TH} \code{TJ}
      \code{TK} \code{TM} \code{TN} \code{TO} \code{TP} \code{TR}
      \code{TT} \code{TV} \code{TW} \code{TZ} \code{UA} \code{UG}
      \code{UK} \code{UM} \code{US} \code{UY} \code{UZ} \code{VA}
      \code{VC} \code{VE} \code{VG} \code{VI} \code{VN} \code{VU}
      \code{WF} \code{WS} \code{YE} \code{YT} \code{ZA} \code{ZM}
      \code{ZR} \code{ZW} \code{ZW-Zimbabwe}} 
   \item{COUNTRY}{A character vector with country name for the
     corresponding code} 
}}
\examples{
data(countries)
head(countries)
}
\keyword{datasets}
