% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_rising_search.R
\name{extract_top_rising}
\alias{extract_top_rising}
\title{Extract top and rising searches}
\usage{
extract_top_rising(data)
}
\arguments{
\item{data}{Data returned by a search query.}
}
\value{
A list with TOP and RISING data (if such data was found).
}
\description{
Extract top and rising searches
}
