% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trends.R
\name{get_interest}
\alias{get_interest}
\title{Retrieve interest over time}
\usage{
get_interest(gtrends)
}
\arguments{
\item{gtrends}{An object of class gtrends created via \code{query_trends()}}
}
\description{
Extract a tibble of interest over time
}
\examples{
\dontrun{
ob <- query_trends("obama")
get_interest(ob)
}

}
