% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trends.R
\name{get_interest_city}
\alias{get_interest_city}
\title{Retrieve interest by city}
\usage{
get_interest_city(trendy)
}
\arguments{
\item{trendy}{An object of class trendy created via \code{trendy()}}
}
\description{
Extract a tibble of interest by city from trendy object
}
\examples{
\dontrun{
ob <- trendy("obama")
get_interest_city(ob)
}

}
