% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trends.R
\name{get_interest_country}
\alias{get_interest_country}
\title{Retrieve interest by country}
\usage{
get_interest_country(trendy)
}
\arguments{
\item{trendy}{An object of class trendy created via \code{trendy()}}
}
\description{
Extract a tibble of interest by country from trendy object
}
\examples{
\dontrun{
ob <- trendy("obama")
get_interest_country(ob)
}

}
