% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trends.R
\name{get_interest_region}
\alias{get_interest_region}
\title{Retrieve interest by region}
\usage{
get_interest_region(trendy)
}
\arguments{
\item{trendy}{An object of class trendy created via \code{trendy()}}
}
\description{
Extract a tibble of interest by region from trendy object
}
\examples{
\dontrun{
ob <- trendy("obama")
get_interest_region(ob)
}
}
