% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trends.R
\name{get_related_queries}
\alias{get_related_queries}
\title{Retrieve related queries}
\usage{
get_related_queries(trendy)
}
\arguments{
\item{trendy}{An object of class trendy created via \code{trendy()}}
}
\value{
A tibble containing related search terms
}
\description{
Extract tibble of related queries from trendy object
}
\examples{
\dontrun{
ob <- trendy("obama")
get_related_queries(ob)
}

}
