% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trends.R
\name{get_related_topics}
\alias{get_related_topics}
\title{Retrieve related topics}
\usage{
get_related_topics(trendy)
}
\arguments{
\item{trendy}{An object of class trendy created via \code{trendy()}}
}
\description{
Extract a tibble of related topics from trendy object
}
\examples{
\dontrun{
ob <- trendy("obama")
get_related_topics(ob)
}

}
