% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtrends.R
\name{getDefaultConnection}
\alias{getDefaultConnection}
\title{Retrieve the default connection object}
\usage{
getDefaultConnection()
}
\value{
An connection handle object as created by \code{\link{gconnect}}
    and stored in the package-local environment.
}
\description{
See the documentation for \code{\link{gconnect}} for the available
options to store default user and password information.
}
\seealso{
\code{\link{gconnect}}
}

