% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtrends.R
\name{gtrends}
\alias{gtrends}
\title{Google Trends Query}
\usage{
gtrends(keyword, geo = "", time = "today+5-y", gprop = c("web", "news",
  "images", "froogle", "youtube"), category = 0)
}
\arguments{
\item{keyword}{A character vector with the actual Google Trends query 
keywords. Multiple keywords are possible using \code{gtrends(c("NHL", 
"NBA", "MLB", "MLS"))}.}

\item{geo}{A character vector denoting geographic regions for the query, 
default to \dQuote{all} for global queries. Multiple regions are possible 
using \code{gtrends("NHL", c("CA", "US"))}.}

\item{time}{The time}

\item{gprop}{A character string defining the Google product for which the 
trend query if preformed. Valid options are \dQuote{} (empty string - web 
search), \dQuote{news}, \dQuote{images}, \dQuote{froogle} and 
\dQuote{youtube}. Default is \dQuote{}.}

\item{category}{A character denoting the category, defaults to \dQuote{0}.}
}
\value{
An object of class \sQuote{gtrends} (basically a list of data frames).
}
\description{
The \code{gtrends} default method performs a Google Trends query for the 
\sQuote{query} argument and session \sQuote{session}. Optional arguments for 
geolocation and category can also be supplied.
}
\section{Categories}{
 The package includes a complete list of categories that 
  can be used to narrow requests. These can be accessed using 
  \code{data("categories")}.
}

\examples{

head(gtrends("NHL"))

head(gtrends(c("NHL", "NFL")))

head(gtrends(c("NHL", "NFL"), geo = c("CA", "US")))

## Sport category (20)
data(categories)
categories[grepl("^Sport", categories$name), ]
head(gtrends(c("NHL", "NFL"), geo = c("CA", "US"), category = 20))

## Playing with time format

head(gtrends(c("NHL", "NFL"), time = "now 1-H")) # last hour
head(gtrends(c("NHL", "NFL"), time = "now 4-H")) # last four hours
head(gtrends(c("NHL", "NFL"), time = "now 1-d")) # last day
head(gtrends(c("NHL", "NFL"), time = "today 1-m")) # last 30 days
head(gtrends(c("NHL", "NFL"), time = "today 3-m")) # last 90 days
head(gtrends(c("NHL", "NFL"), time = "today 12-m")) # last 12 months
head(gtrends(c("NHL", "NFL"), time = "today+5-y")) # last five years (default)
head(gtrends(c("NHL", "NFL"), time = "all")) # since 2004

## Custom date format

head(gtrends(c("NHL", "NFL"), time = "2010-01-01 2010-04-03")) 

## Search from various Google's services

head(gtrends(c("NHL", "NFL"), gprop = "news")$interest_over_time)
head(gtrends(c("NHL", "NFL"), gprop = "youtube")$interest_over_time)

}
