% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgraph.R
\name{kgraph}
\alias{kgraph}
\title{kgraph}
\usage{
kgraph(keyword, token, ids = "", hl = "", types = "", prefix = FALSE,
  limit = 10)
}
\arguments{
\item{keyword}{A character vector with the actual Google Knowledge Graph
Search query keyword. Note that only one keyword allowed a time.}

\item{token}{A character string containing your Google API access token.
Details about how to obtain your access token can be found here
(\url{https://developers.google.com/knowledge-graph/how-tos/authorizing}).}

\item{ids}{A vector of entity IDs to search for in the Knowledge Graph.}

\item{hl}{A string specifying the ISO 639 language code (ex.: \dQuote{en} or 
\dQuote{fr}).}

\item{types}{A vector of character strings which restricts the returned 
entities. For example, you can specify Person to restrict the results to 
entities representing people. If multiple types are specified, returned 
entities will contain one or more of these types. Full list of schemas
types is availabe (\url{http://schema.org/docs/full.html})}

\item{prefix}{If set to \dQuote{TRUE} prefix (initial substring) match
against names and aliases of entities is allowed. For example, a prefix
Jung will match entities and aliases such as Jung, Jungle, and Jung-ho
Kang. Default is \dQuote{FALSE}.}

\item{limit}{Sets the maximum of returned entities per call. Default is
\dQuote{10}, maximum allowed is \dQuote{20}}
}
\value{
Returns an object of class \sQuote{kgraph}. This is a list containing
  the entities returned from the call sorted in an ascending order relative
  to the relevance score.
}
\description{
Obtains Google Knowledge Graph Search entities for a given keyword.
}
\details{
\code{kgraph} can be used to obtain Google Knowledge Graph Search entities
(\url{https://www.google.com/intl/es419/insidesearch/features/search/knowledge.html})
for any given keyword. The obtained \dQuote{kg-id} can then be used together
with the \code{gtrends} function to obtain Google Trends information based on
the Topic Search. This is a set of combined search queries providing the 
overall search interest on the topic. For example the topic search query for 
\dQuote{London the capital of England}, will not only cover searches for the
keyword London but also for \dQuote{Hotel in London}  obtaining  to be used
as \code{gtrends} Google Topic search function in Google Trends.
}
\examples{
kg <- kgraph("Call of Duty 2", "YOUR TOKEN", types = "VideoGame")
# get google trends for the first entity
topicsearch <- gtrends(kgs$entities[[1]]$id, time = "all")
        
}
\author{
Oliver Schaer, \email{info@oliverschaer.ch}
}
