% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtrends.R
\name{plot.gtrends}
\alias{plot.gtrends}
\title{Plot Google Trends interest over time}
\usage{
\method{plot}{gtrends}(x, ...)
}
\arguments{
\item{x}{A \code{\link{gtrends}} object.}

\item{...}{Additional parameters passed on in method dispatch. Currently not
used.}
}
\value{
A ggplot2 object is returned silently.
}
\description{
Plot Google Trends interest over time
}
\examples{
\dontrun{
res <- gtrends("nhl", geo = c("CA", "US"))
plot(res)
}
}
