% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{print.gtrends}
\alias{print.gtrends}
\title{Create print method for trendy object}
\usage{
\method{print}{gtrends}(x, ...)
}
\arguments{
\item{x}{trendy object}
}
\description{
Create print method for trendy object
}
