% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtrends.R
\name{which_type}
\alias{which_type}
\title{Guess block type}
\usage{
which_type(block)
}
\arguments{
\item{block}{A block of information returned by Google Trends.}
}
\value{
The block name.
}
\description{
Guess block type
}
\details{
Utility function that tries to determine which kind of block was
  returned by Google Trends (regions, metro, city, etc.).
}

