% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\docType{data}
\name{countries}
\alias{countries}
\title{Word countries ISO code.}
\format{A data frame with 244 rows and 2 variables}
\usage{
data("countries")
}
\description{
\itemize{
  \item CODE  ISO 3 characters codes
  \item COUNTRY Full countries names
}
}
\keyword{datasets}

