% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\docType{data}
\name{countries}
\alias{countries}
\title{Word countries ISO code.}
\format{A data frame with 4707 rows and 4 variables}
\usage{
data("countries")
}
\description{
\itemize{
  \item code  ISO 3 characters codes
  \item country Full countries names
  \item subcode Countries subcodes
  \item subname Countries subname
}
}
\references{
\url{http://www.unece.org/cefact/codesfortrade/codes_index.html}
}
\keyword{datasets}

