% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\docType{data}
\name{countries}
\alias{countries}
\title{Word countries ISO code.}
\format{A data frame with 89743 rows and 3 variables}
\usage{
data("countries")
}
\description{
\itemize{
  \item country_code  Two-digits country codes
  \item description Descption of the location
  \item sub_code ISO3166-2 country codes
}
}
\references{
\url{http://www.unece.org/cefact/codesfortrade/codes_index.html}
}
\keyword{datasets}

