% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtrends.R
\name{gconnect}
\alias{.getDefaultConnection}
\alias{gconnect}
\title{Connect to Google account}
\usage{
gconnect(usr = NULL, psw = NULL, verbose = FALSE)

.getDefaultConnection()
}
\arguments{
\item{usr}{Username (ex.: yourmail@gmail.com)}

\item{psw}{Account password}

\item{verbose}{Logical for displaying additional information}
}
\value{
A libcurl handle is returned (invisibly).
}
\description{
The resulting connection object is also stored in the package-local
environment from which the (internal) helper function
\code{.getDefaultConnection()} retrieves it as needed.
}
\examples{
\dontrun{
ch <- gconnect("usr@gmail.com", "psw")
}
}

