% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtrends.R
\name{gconnect}
\alias{gconnect}
\title{Connect to Google account}
\usage{
gconnect(usr = NULL, psw = NULL, verbose = FALSE)
}
\arguments{
\item{usr}{Username (ex.: yourmail@gmail.com)}

\item{psw}{Account password}

\item{verbose}{Logical for displaying additional information}
}
\value{
A libcurl handle
}
\description{
Connect to Google account
}
\examples{
\dontrun{
gconnect("myouremail@gmail.com", "mysuperpassword")
}
}

