% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trends.R
\name{get_interest}
\alias{get_interest}
\title{Retrieve interest over time}
\usage{
get_interest(trendy)
}
\arguments{
\item{trendy}{An object of class trendy created via \code{trendy()}}
}
\description{
Extract a tibble of interest over time
}
\examples{
\dontrun{
ob <- trendy("obama")
get_interest(ob)
}

}
