% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trends.R
\name{get_interest_city}
\alias{get_interest_city}
\title{Retrieve interest by city}
\usage{
get_interest_city(gtrends)
}
\arguments{
\item{gtrends}{An object of class gtrends created via \code{query_trends()}}
}
\description{
Extract a tibble of interest by city from gtrends object
}
\examples{
\dontrun{
ob <- query_trends("obama")
get_interest_city(ob)
}

}
