% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trends.R
\name{get_interest_country}
\alias{get_interest_country}
\title{Retrieve interest by country}
\usage{
get_interest_country(gtrends)
}
\arguments{
\item{gtrends}{An object of class gtrends created via \code{query_trends()}}
}
\description{
Extract a tibble of interest by country from gtrends object
}
\examples{
\dontrun{
ob <- query_trends("obama")
get_interest_country(ob)
}

}
