% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trends.R
\name{get_interest_dma}
\alias{get_interest_dma}
\title{Retrieve interest by DMA}
\usage{
get_interest_dma(gtrends)
}
\arguments{
\item{gtrends}{An object of class gtrends created via \code{query_trends()}}
}
\description{
Extract a tibble of interest by DMA from gtrends object
}
\examples{
\dontrun{
ob <- query_trends("obama")
get_interest_dma(ob)
}

}
