% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trends.R
\name{get_related_queries}
\alias{get_related_queries}
\title{Retrieve related queries}
\usage{
get_related_queries(gtrends)
}
\arguments{
\item{gtrends}{An object of class gtrends created via \code{query_trends()}}
}
\value{
A tibble containing related search terms
}
\description{
Extract tibble of related queries from gtrends object
}
\examples{
\dontrun{
ob <- query_trends("obama")
get_related_queries(ob)
}

}
