% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trends.R
\name{get_related_topics}
\alias{get_related_topics}
\title{Retrieve related topics}
\usage{
get_related_topics(gtrends)
}
\arguments{
\item{gtrends}{An object of class gtrends created via \code{query_trends()}}
}
\description{
Extract a tibble of related topics from gtrends object
}
\examples{
\dontrun{
ob <- query_trends("obama")
get_related_topics(ob)
}

}
