% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_trends.R
\name{print.gtrends}
\alias{print.gtrends}
\title{Print gtrends object}
\usage{
\method{print}{gtrends}(x, ...)
}
\arguments{
\item{x}{gtrends object}

\item{...}{Additional parameters passed on in method dispatch. Currently not
used.}
}
\description{
Print gtrends object
}
